+++ lib/joystick/joystick_controller.dart
@@ -2,10 +2,10 @@
 
 import 'package:bonfire/base/game_component.dart';
 import 'package:bonfire/joystick/joystick.dart';
+import 'package:bonfire/mixins/keyboard_listener.dart';
 import 'package:bonfire/mixins/pointer_detector.dart';
 import 'package:bonfire/util/priority_layer.dart';
 import 'package:flame/components.dart';
-import 'package:flutter/widgets.dart';
 
 enum JoystickMoveDirectional {
   MOVE_UP,
@@ -54,12 +54,10 @@ mixin JoystickListener {
 }
 
 abstract class JoystickController extends GameComponent
-    with PointerDetectorHandler {
+    with PointerDetectorHandler, KeyboardEventListener {
   final List<JoystickListener> _observers = [];
 
   KeyboardConfig keyboardConfig = KeyboardConfig(enable: false);
-
-  KeyEventResult onKeyboard(RawKeyEvent event) => KeyEventResult.handled;
 
   void joystickChangeDirectional(JoystickDirectionalEvent event) {
     for (var o in _observers) {